refactor(EditPolicy): RHICOMPL-1696 split to EditPolicyForm
